+++ exchange/errors.py
@@ -1,0 +1,6 @@
+class ExchangeError(Exception):
+    pass
+
+
+class ExchangeConfigError(ExchangeError):
+    pass