+++ exchange/urls.py
@@ -3,6 +3,7 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path('', views.convert, name='convert'),
-    path('', views.rates, name='rates'),
+    # TODO
+    # path('', views.convert, name='convert'),
+    # path('', views.rates, name='rates'),
 ]